Refactor and fix bugs
