+++ assets/css/availability-form.scss
@@ -1,5 +1,4 @@
-table.availability-table {
-  text-align: center;
+table.availability-form-table {
   position: relative;
 
   label.day-title {
@@ -12,8 +11,7 @@ table.availability-table {
     min-width: 80px;
   }
 
-  td.clickable-table-box {
-    cursor: pointer;
+  td.slot-box {
     min-width: 85px;
 
     &:hover {
@@ -28,24 +26,8 @@ table.availability-table {
       }
     }
 
-    &.highlight {
-      box-shadow: inset 0 0 0 4px crimson;
-    }
-
-    &.disabled {
-      cursor: default;
-    }
-
     &.booked {
       background-color: #67a2d3 !important;
-    }
-
-    &.locked {
-      background-color: #7a7a7a !important;
-    }
-
-    input[type=checkbox] {
-      cursor: pointer;
     }
   }
