+++ assets/css/availability-table.scss
@@ -1,0 +1,50 @@
+table.availability-table {
+  text-align: center;
+
+  td.slot-box {
+    vertical-align: middle;
+    cursor: pointer;
+
+    input[type='checkbox'] {
+      cursor: pointer;
+    }
+
+    &:hover {
+      opacity: 0.8;
+    }
+
+    &.disabled {
+      cursor: default;
+    }
+  }
+
+  &.shift-pressed {
+    td.prev-clicked {
+      opacity: 1;
+      box-shadow: inset 0 0 0 4px crimson;
+    }
+  }
+}
+
+table.availability-table,
+.planning-actions-container {
+  .unknown {
+    background-color: #ff767f;
+    color: white;
+  }
+
+  .available {
+    background-color: #76bd79;
+    color: white;
+  }
+
+  .booked {
+    background-color: #67a2d3;
+    color: white;
+  }
+
+  .locked {
+    background-color: #7a7a7a;
+    color: white;
+  }
+}
