+++ assets/css/planning.scss
@@ -21,7 +21,6 @@ $tableBoxSize: 40px;
 
 .planning-container {
   .planning {
-    text-align: center;
     margin-bottom: 40px;
 
     .badge {
@@ -112,21 +111,9 @@ $tableBoxSize: 40px;
     td.slot-box {
       width: $tableBoxSize;
       height: $tableBoxSize;
-      vertical-align: middle;
-      cursor: pointer;
 
       input[type='checkbox'] {
-        cursor: pointer;
         visibility: hidden;
-      }
-
-      &:hover {
-        opacity: 0.8;
-      }
-
-      &.highlight {
-        opacity: 1;
-        box-shadow: inset 0 0 0 4px crimson;
       }
 
       &.checked {
@@ -134,10 +121,6 @@ $tableBoxSize: 40px;
 
         input[type='checkbox'] {
           visibility: visible;
-        }
-
-        &.highlight {
-          opacity: 1;
         }
       }
     }
@@ -148,26 +131,6 @@ $tableBoxSize: 40px;
 .planning-actions-container {
   .item-data {
     display: none;
-  }
-
-  .unknown {
-    background-color: #ff767f;
-    color: white;
-  }
-
-  .available {
-    background-color: #76bd79;
-    color: white;
-  }
-
-  .booked {
-    background-color: #67a2d3;
-    color: white;
-  }
-
-  .locked {
-    background-color: #7a7a7a;
-    color: white;
   }
 }
