+++ assets/js/planning-update.js
@@ -1,51 +1,6 @@
 const $ = require('jquery');
 require('bootstrap');
 
-function colorTableBox($tableBox) {
-  const isChecked = $tableBox.find('input:checkbox').prop('checked');
-  $tableBox.toggleClass('checked', isChecked);
-}
-
-function selectTableBox($tableBox) {
-  if (!$tableBox) {
-    return;
-  }
-
-  const $checkbox = $tableBox.find('input:checkbox');
-  if ($checkbox.prop('disabled')) {
-    return;
-  }
-
-  $checkbox.prop('checked', !$checkbox.prop('checked'));
-  colorTableBox($tableBox);
-}
-
-function handleSlotBoxClick(event) {
-  event.stopImmediatePropagation();
-
-  const $slotBox = event.target === 'input' ? $(this).closest('.slot-box') : $(this);
-  selectTableBox($slotBox);
-
-  const $planning = $('.planning');
-  const $lastClickedTd = $planning.find('.prev-clicked');
-
-  if ($lastClickedTd && event.shiftKey) {
-    handleShiftClick($planning, $slotBox, $lastClickedTd);
-    $lastClickedTd.removeClass('prev-clicked');
-    $lastClickedTd.removeClass('highlight');
-  }
-
-  $slotBox.addClass('prev-clicked');
-}
-
-function handleShiftInput(event) {
-  if (!event.repeat) {
-    $('.planning')
-      .find('.prev-clicked')
-      .toggleClass('highlight', event.type === 'keydown');
-  }
-}
-
 function triggerUpdate(url, newStatus, $planning, $modal) {
   const payload = generatePayload($planning);
 
@@ -148,76 +103,8 @@ function checkLastUpdate(forceUpdate) {
   });
 }
 
-function handleShiftClick($planning, $currentClickedTd, $lastClickedTd) {
-  window.getSelection().removeAllRanges();
-  const isChecked = $currentClickedTd.hasClass('checked');
-
-  const minTbodyIndex = Math.min($lastClickedTd.closest('tbody').index(), $currentClickedTd.closest('tbody').index());
-  const maxTbodyIndex = Math.max($lastClickedTd.closest('tbody').index(), $currentClickedTd.closest('tbody').index());
-  const tdFrom = Math.min(Date.parse($lastClickedTd.data('from')), Date.parse($currentClickedTd.data('from')));
-  const tdTo = Math.max(Date.parse($lastClickedTd.data('to')), Date.parse($currentClickedTd.data('to')));
-
-  // default case : maxTbodyIndex === minTbodyIndex
-  let minTrIndex = Math.min($lastClickedTd.closest('tr').index(), $currentClickedTd.closest('tr').index());
-  let maxTrIndex = Math.max($lastClickedTd.closest('tr').index(), $currentClickedTd.closest('tr').index());
-  if (maxTbodyIndex !== minTbodyIndex) {
-    if ($lastClickedTd.closest('tbody').index() === minTbodyIndex) {
-      minTrIndex = $lastClickedTd.closest('tr').index();
-    } else if ($currentClickedTd.closest('tbody').index() === minTbodyIndex) {
-      minTrIndex = $currentClickedTd.closest('tr').index();
-    }
-
-    if ($lastClickedTd.closest('tbody').index() === maxTbodyIndex) {
-      maxTrIndex = $lastClickedTd.closest('tr').index();
-    } else if ($currentClickedTd.closest('tbody').index() === maxTbodyIndex) {
-      maxTrIndex = $currentClickedTd.closest('tr').index();
-    }
-  }
-
-  const handleTbody = function (i, tbody) {
-    const currentTbodyIndex = $(tbody).index();
-
-    $(tbody)
-      .find('tr')
-      .filter((i, tr) => {
-        const trIndex = $(tr).index();
-        if (minTbodyIndex !== maxTbodyIndex) {
-          if (currentTbodyIndex === minTbodyIndex) {
-            return trIndex >= minTrIndex;
-          } else if (currentTbodyIndex === maxTbodyIndex) {
-            return trIndex <= maxTrIndex;
-          }
-
-          return true;
-        }
-
-        return trIndex >= minTrIndex && trIndex <= maxTrIndex;
-      })
-      .each(handleTr);
-  };
-
-  const handleTr = function (i, tr) {
-    $(tr)
-      .find('td')
-      .filter((i, td) => Date.parse($(td).data('from')) >= tdFrom && Date.parse($(td).data('to')) <= tdTo)
-      .each(function () {
-        if (isChecked !== $(this).hasClass('checked')) {
-          selectTableBox($(this));
-        }
-      });
-  };
-
-  $planning
-    .find('tbody.item-rows')
-    .filter((index, currentTbody) => $(currentTbody).index() >= minTbodyIndex && $(currentTbody).index() <= maxTbodyIndex)
-    .each(handleTbody);
-}
-
 $(document).ready(function () {
   const $planning = $('.planning');
-
-  $planning.on('click', '.slot-box', handleSlotBoxClick);
-  $(document).on('keydown keyup', handleShiftInput);
 
   let urlParams = new URLSearchParams(window.location.search);
   if (urlParams.has('scrollTop')) {
