+++ templates/availability/_table.html.twig
@@ -29,7 +29,7 @@
 
 {{ form_start(form) }}
 <div class="table-responsive">
-    <table class="table availability-table">
+    <table class="table availability-table availability-form-table">
         {% set choices = form.availabilityDomains.vars.data %}
         {% set nbChoicesPerDay = choices|length/7 %}
         <thead>
@@ -50,7 +50,7 @@
             <tr>
                 <td class="hour-title">{{ choices[i].date|date('H') }}-{{ choices[i+1].date|date('H') }}h</td>
                 {% for j in 0..6 %}
-                    <td class="clickable-table-box" data-day="{{ j }}" data-from="{{ choices[i+1].date|date('U') }}" data-to="{{ choices[i+1].date|date('U') }}" data-status="{{ '' }}">
+                    <td class="slot-box" data-day="{{ j }}" data-from="{{ choices[i+1].date|date('U') }}" data-to="{{ choices[i+1].date|date('U') }}" data-status="{{ '' }}">
                         {{ form_widget(form.availabilityDomains.children[j*nbChoicesPerDay+i].tick) }}
                     </td>
                 {% endfor %}
