+++ templates/organization/commissionable_asset/availability.html.twig
@@ -5,6 +5,7 @@
 {% block javascripts %}
     {{ parent() }}
     {{ encore_entry_script_tags('availability-form') }}
+    {{ encore_entry_script_tags('availability-table') }}
 {% endblock %}
 
 {% block body %}
