+++ templates/organization/planning/_results.html.twig
@@ -37,7 +37,7 @@
 </div>
 
 <div class="planning-container">
-    <table class="planning table table-sm table-bordered" style="display: none;">
+    <table class="planning table table-sm table-bordered availability-table" data-availability-mode="planning" style="display: none;">
         <thead>
         <tr class="days">
             <th rowspan="2"></th>
