+++ templates/organization/planning/planning.html.twig
@@ -4,6 +4,7 @@
     {{ parent() }}
     {{ encore_entry_script_tags('planning') }}
     {{ encore_entry_script_tags('planning-update') }}
+    {{ encore_entry_script_tags('availability-table') }}
 {% endblock %}
 
 {% block title %}Planning{% endblock %}
